Fix aileronTrim

+++ MsfsPlugin/MsfsPlugin/encoder/AltitudeAPEncoder.cs
@@ -9,6 +9,6 @@
         public AltitudeAPEncoder() : base("Alt", "Autopilot altitude encoder", "AP", true, -10000, 99900, 100) { }
         protected override void RunCommand(String actionParameter) => MsfsData.Instance.CurrentAPAltitude = MsfsData.Instance.CurrentAltitude;
         protected override String GetDisplayValue() => "[" + MsfsData.Instance.CurrentAPAltitudeState + "]\n" + MsfsData.Instance.CurrentAltitude;
-        protected override void SetValue(Int32 newValue) => MsfsData.Instance.CurrentAPAltitude = newValue;
+        protected override void SetValue(Int64 newValue) => MsfsData.Instance.CurrentAPAltitude = (Int32)newValue;
     }
 }
