+++ MsfsPlugin/MsfsPlugin/encoder/HeadignAPEncoder.cs
@@ -9,6 +9,6 @@
         public HeadignAPEncoder() : base("Head", "Autopilot heading encoder", "AP", true, 0, 360, 1) { }
         protected override void RunCommand(String actionParameter) => MsfsData.Instance.CurrentAPHeading = MsfsData.Instance.CurrentHeading;
         protected override String GetDisplayValue() => "[" + MsfsData.Instance.CurrentAPHeadingState + "]\n" + MsfsData.Instance.CurrentHeading;
-        protected override void SetValue(Int32 newValue) => MsfsData.Instance.CurrentAPHeading = newValue;
+        protected override void SetValue(Int64 newValue) => MsfsData.Instance.CurrentAPHeading = (Int32)newValue;
     }
 }
