+++ MsfsPlugin/MsfsPlugin/encoder/RudderTrimEncoder.cs
@@ -8,7 +8,7 @@
     {
         public RudderTrimEncoder() : base("Rudder Trim", "Rudder trim encoder", "Nav", true, -100, 100, 1) { }
         protected override void RunCommand(String actionParameter) => this.SetValue(0);
-        protected override Int32 GetValue() => MsfsData.Instance.CurrentRudderTrim;
-        protected override void SetValue(Int32 newValue) => MsfsData.Instance.CurrentRudderTrim = newValue;
+        protected override Int64 GetValue() => MsfsData.Instance.CurrentRudderTrim;
+        protected override void SetValue(Int64 newValue) => MsfsData.Instance.CurrentRudderTrim = (Int32)newValue;
     }
 }
