+++ MsfsPlugin/MsfsPlugin/encoder/ThrottleEncoder.cs
@@ -7,11 +7,11 @@
     {
         public ThrottleEncoder() : base("Throttle", "Current throttle", "Nav", true, -100, 100, 1) { }
         protected override void RunCommand(String actionParameter) => this.SetValue(MsfsData.Instance.ThrottleLowerFromMSFS < 0 ? -100 : 0);
-        protected override Int32 GetValue()
+        protected override Int64 GetValue()
         {
             this.min = MsfsData.Instance.ThrottleLowerFromMSFS < 0 ? -100 : 0;
             return MsfsData.Instance.CurrentThrottle;
         }
-        protected override void SetValue(Int32 newValue) => MsfsData.Instance.CurrentThrottle = newValue;
+        protected override void SetValue(Int64 newValue) => MsfsData.Instance.CurrentThrottle = (Int32)newValue;
     }
 }
